Ripley's K Function

Computes Ripley's K function to assess spatial clustering or dispersion at multiple scales.

Usage

RipleysK(object, r_seq = NULL, target = NULL, correction = c("none", "border"))

Arguments

object

An SpatialCellData-class object.

r_seq

Numeric vector of radii, or NULL for automatic.

target

Character or NULL. Restrict to a phenotype.

correction

Character. "none" (default) or "border".

Value

A data frame with columns r, K, L, and expected.

Examples

counts <- matrix(rnorm(100), nrow = 50,
                 dimnames = list(NULL, c("CD3", "CD8")))
coords <- data.frame(x = runif(50, 0, 200), y = runif(50, 0, 200))
obj <- CreateSpatialObject(counts, coords)
RipleysK(obj)
#>             r          K          L    expected
#> 1   0.0000000    0.00000  0.0000000    0.000000
#> 2   0.9691923    0.00000 -0.9691923    2.951004
#> 3   1.9383845    0.00000 -1.9383845   11.804015
#> 4   2.9075768   29.65506  0.1648001   26.559033
#> 5   3.8767690   59.31013  0.4682280   47.216058
#> 6   4.8459613   59.31013 -0.5009642   73.775091
#> 7   5.8151536   59.31013 -1.4701565  106.236132
#> 8   6.7843458   59.31013 -2.4393487  144.599179
#> 9   7.7535381   59.31013 -3.4085410  188.864234
#> 10  8.7227303  118.62025 -2.5779765  239.031296
#> 11  9.6919226  177.93038 -2.1661669  295.100366
#> 12 10.6611149  207.58544 -2.5323697  357.071442
#> 13 11.6303071  266.89557 -2.4131764  424.944526
#> 14 12.5994994  296.55063 -2.8837906  498.719618
#> 15 13.5686916  355.86076 -2.9256659  578.396717
#> 16 14.5378839  415.17089 -3.0421022  663.975823
#> 17 15.5070762  504.13608 -2.8393417  755.456936
#> 18 16.4762684  563.44620 -3.0840880  852.840056
#> 19 17.4454607  622.75633 -3.3660610  956.125184
#> 20 18.4146529  741.37658 -3.0527685 1065.312320
#> 21 19.3838452  919.30696 -2.2775746 1180.401462
#> 22 20.3530375  978.61709 -2.7035759 1301.392612
#> 23 21.3222297 1067.58228 -2.8879683 1428.285769
#> 24 22.2914220 1186.20253 -2.8600043 1561.080934
#> 25 23.2606143 1275.16772 -3.1136916 1699.778106
#> 26 24.2298065 1393.78797 -3.1666517 1844.377285
#> 27 25.1989988 1512.40823 -3.2578390 1994.878471
#> 28 26.1681910 1601.37341 -3.5909239 2151.281665
#> 29 27.1373833 1631.02848 -4.3520264 2313.586866
#> 30 28.1065756 1838.61392 -3.9146557 2481.794074
#> 31 29.0757678 1927.57911 -4.3054734 2655.903290
#> 32 30.0449601 2194.47468 -3.6153747 2835.914513
#> 33 31.0141523 2313.09493 -3.8796543 3021.827743
#> 34 31.9833446 2520.68038 -3.6574292 3213.642981
#> 35 32.9525369 2728.26582 -3.4833329 3411.360226
#> 36 33.9217291 2906.19620 -3.5067496 3614.979478
#> 37 34.8909214 3143.43670 -3.2588652 3824.500738
#> 38 35.8601136 3291.71202 -3.4906144 4039.924005
#> 39 36.8293059 3469.64240 -3.5964686 4261.249279
#> 40 37.7984982 3706.88290 -3.4482802 4488.476560
#> 41 38.7676904 3795.84809 -4.0077138 4721.605849
#> 42 39.7368827 4003.43354 -4.0390890 4960.637145
#> 43 40.7060749 4299.98417 -3.7097579 5205.570449
#> 44 41.6752672 4359.29430 -4.4246770 5456.405759
#> 45 42.6444595 4507.56961 -4.7656531 5713.143077
#> 46 43.6136517 4715.15505 -4.8724531 5975.782403
#> 47 44.5828440 4833.77531 -5.3573616 6244.323735
#> 48 45.5520362 4982.05062 -5.7294802 6518.767075
#> 49 46.5212285 5219.29113 -5.7615430 6799.112423
#> 50 47.4904208 5397.22151 -6.0417903 7085.359777